Refactor profile edition
